fix: imageをbinaryで返すように修正
